terraform/functions/process_audio/requirements.txt
google-cloud-speech>=2.21.0
google-cloud-translate>=3.11.1
google-cloud-firestore>=2.11.0
